requirements.txt
sqlalchemy~=1.4.9
python-telegram-bot==20.0a0
aiohttp~=3.8.1
asyncio~=3.4.3
python-dotenv~=0.20.0
psycopg2-binary